+++ packages/visually-hidden/examples/default.tsx
@@ -1,0 +1,14 @@
+import { VisuallyHidden, type VisuallyHiddenProps } from '@bento/visually-hidden';
+/* v8 ignore next */
+import React from 'react';
+
+export function DefaultExample(args: VisuallyHiddenProps) {
+  return (
+    <div>
+      <VisuallyHidden isFocusable {...args}>
+        <button>Skip to main content</button>
+      </VisuallyHidden>
+      Hit tab to focus the visually hidden button.
+    </div>
+  );
+}