+++ scripts/test/vitest.config.ts
@@ -1,0 +1,18 @@
+import { dirname, join } from 'node:path';
+import { fileURLToPath } from 'node:url';
+
+import react from '@vitejs/plugin-react';
+import tsconfigPaths from 'vite-tsconfig-paths';
+import { defineConfig } from 'vitest/config';
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
+
+export default defineConfig({
+  plugins: [react(), tsconfigPaths()],
+  test: {
+    globals: true,
+    name: 'Scripts Tests',
+    include: [join(__dirname, '**/*.node.test.{ts,tsx}')],
+    environment: 'node'
+  }
+});