+++ packages/checkbox/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @bento/checkbox
 
+## 0.1.3
+
+### Patch Changes
+
+- Adds documentation to all packages; no API or functionality changes. ([#23](https://github.com/godaddy/bento/pull/23) by @kawikabader)
+
+<details>
+<summary>Updated dependencies</summary>
+
+- @bento/control@0.1.2
+- @bento/use-data-attributes@0.1.1
+- @bento/use-props@0.1.1
+- @bento/slots@0.1.3
+- @bento/icon@0.1.2
+</details>
+
 ## 0.1.2
 
 ### Patch Changes