+++ packages/use-data-attributes/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @bento/use-data-attributes
 
+## 0.1.1
+
+### Patch Changes
+
+- Adds documentation to all packages; no API or functionality changes. ([#23](https://github.com/godaddy/bento/pull/23) by @kawikabader)
+
+<details>
+<summary>Updated dependencies</summary>
+
+- @bento/to-attribute-value@0.1.1
+</details>
+
 ## 0.1.0
 
 ### Minor Changes